# Conflicts:
#	main.py

+++ main.py
@@ -4,7 +4,7 @@
 user_input_list = []
 for i in range(0, 10):
     user_input = input(f"숫자를 입력해주세요({i}/10)")
-    user_input = int(user_input)
+    user_input = float(user_input)
     user_input_list.append(user_input)
